scripts/requirements.txt
PyGithub==1.55
openai==1.99.1
